refactor: fps 更新処理を別の関数に切り出し
